add helper 1. add helper class hidden table 2. add helper class watch table
